requirements.txt
fastapi
uvicorn
apscheduler
sqlalchemy
psycopg2
python-dotenv